+++ PropDefiniteKB.py
@@ -1,0 +1,18 @@
+from PropKB import *
+
+
+class PropDefiniteKB(PropKB):
+    """A KB of propositional definite clauses."""
+
+    def tell(self, sentence):
+        """Add a definite clause to this KB."""
+        assert is_definite_clause(sentence), "Must be definite clause"
+        self.clauses.append(sentence)
+
+    def retract(self, sentence):
+        self.clauses.remove(sentence)
+
+    def clauses_with_premise(self, p):
+        """Return a list of the clauses in KB that have p in their premise.
+        This could be cached away for O(1) speed, but we'll recompute it."""
+        return [c for c in self.clauses if c.op == '==>' and p in conjuncts(c.args[0])]
