+++ bc.py
@@ -1,0 +1,10 @@
+from collections import defaultdict
+from cnf import *
+from expr import *
+
+symbols = []
+
+
+def pl_bc_entails(kb, q):
+
+    return False
