+++ fc.py
@@ -1,10 +1,7 @@
 from collections import defaultdict
-from data_convertor import *
+from cnf import *
 from expr import *
-from KB import *
-from PropKB import *
 
-file_name, p, q = data_convertor()
 symbols = []
 
 
@@ -30,25 +27,3 @@ def pl_fc_entails(kb, q):
                 if count[c] == 0:
                     agenda.append(c.args[1])
     return False
-
-
-class PropDefiniteKB(PropKB):
-    """A KB of propositional definite clauses."""
-
-    def tell(self, sentence):
-        """Add a definite clause to this KB."""
-        assert is_definite_clause(sentence), "Must be definite clause"
-        self.clauses.append(sentence)
-
-    def ask_generator(self, query):
-        """Yield the empty substitution if KB implies query; else nothing."""
-        if pl_fc_entails(self.clauses, query):
-            yield {}
-
-    def retract(self, sentence):
-        self.clauses.remove(sentence)
-
-    def clauses_with_premise(self, p):
-        """Return a list of the clauses in KB that have p in their premise.
-        This could be cached away for O(1) speed, but we'll recompute it."""
-        return [c for c in self.clauses if c.op == '==>' and p in conjuncts(c.args[0])]
